Items can be activated by pressing "a"
* Player should stand next to the item and press "a" and then a
direction key in order to activate the object. So far doors are the only
items in game.
* lib/obj contain game objects, items in particular at certain moment.
All items should extend ItemTemplate class and act as an interface.
* Added new waiting_for_action_arg flag (long name, did not come up with
shorter, but still descriptive version); if flag is 1 - argument from
next event will be appended to current action and sent if possible. If
evt_arg value is 0 - flag is reset and no data is sent.